Improve certificate info
It's unclear why this extra file is checked on Linux since it was not
explained when originally introduced.

In addition to that, this may hide real issues when this additional file
does not exist by the actual problem is with the value in
`ENV["SSL_CERT_FILE"]`, for example, because of an unexpanded tilde.